+++ src/main/java/cc/ryanc/halo/model/freemarker/tag/CommonTagDirective.java
@@ -1,8 +1,6 @@
-package cc.ryanc.halo.model.tag;
+package cc.ryanc.halo.model.freemarker.tag;
 
-import cc.ryanc.halo.model.enums.PostStatusEnum;
-import cc.ryanc.halo.model.enums.PostTypeEnum;
-import cc.ryanc.halo.service.PostService;
+import cc.ryanc.halo.service.*;
 import freemarker.core.Environment;
 import freemarker.template.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,12 +18,24 @@ import java.util.Map;
  * @date : 2018/4/26
  */
 @Component
-public class ArticleTagDirective implements TemplateDirectiveModel {
+public class CommonTagDirective implements TemplateDirectiveModel {
 
     private static final String METHOD_KEY = "method";
 
     @Autowired
-    private PostService postService;
+    private MenuService menuService;
+
+    @Autowired
+    private CategoryService categoryService;
+
+    @Autowired
+    private CommentService commentService;
+
+    @Autowired
+    private TagService tagService;
+
+    @Autowired
+    private LinkService linkService;
 
     @Override
     public void execute(Environment environment, Map map, TemplateModel[] templateModels, TemplateDirectiveBody templateDirectiveBody) throws TemplateException, IOException {
@@ -33,17 +43,20 @@ public class ArticleTagDirective implements TemplateDirectiveModel {
         if (map.containsKey(METHOD_KEY)) {
             String method = map.get(METHOD_KEY).toString();
             switch (method) {
-                case "postsCount":
-                    environment.setVariable("postsCount", builder.build().wrap(postService.findPostByStatus(PostStatusEnum.PUBLISHED.getCode(), PostTypeEnum.POST_TYPE_POST.getDesc()).size()));
+                case "menus":
+                    environment.setVariable("menus", builder.build().wrap(menuService.findAll()));
                     break;
-                case "archives":
-                    environment.setVariable("archives", builder.build().wrap(postService.findPostGroupByYearAndMonth()));
+                case "categories":
+                    environment.setVariable("categories", builder.build().wrap(categoryService.findAll()));
                     break;
-                case "archivesLess":
-                    environment.setVariable("archivesLess", builder.build().wrap(postService.findPostGroupByYear()));
+                case "tags":
+                    environment.setVariable("tags", builder.build().wrap(tagService.findAll()));
                     break;
-                case "hotPosts":
-                    environment.setVariable("hotPosts", builder.build().wrap(postService.hotPosts()));
+                case "links":
+                    environment.setVariable("links", builder.build().wrap(linkService.findAll()));
+                    break;
+                case "newComments":
+                    environment.setVariable("newComments", builder.build().wrap(commentService.findAll(1)));
                     break;
                 default:
                     break;
